test: add normalizeBands tests
