+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "preview": "vite preview --port 5173"
+    "preview": "vite preview --port 5173",
+    "test": "node --test"
   },
   "dependencies": {
     "d3": "7.9.0",
@@ -19,5 +20,3 @@
     "vite": "5.4.6"
   }
 }
-
-
